[Messages][HaTS] Add tab observer for HaTS message
This CL includes updates to the HaTS message implementation to keep in line with the existing infobar behavior on events associated with the tab where the prompt is displayed.

The primary issue that is being targeted here is that the message prompt currently shows up even after the survey has expired. To avoid this, the message will be dismissed when the original tab where the message was displayed is hidden for any reason.

Bug: 1254789
Change-Id: Ida34c9f4c2c3e0a428468b5b861fceaa9053d0d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198140
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Pavel Yatsuk <pavely@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927964}