lacros: Add ash conditional for chrome app integration.
This CL has no functional effect.

This CL gates chrome app integration for lacros behind a conditional in
ash. Ash propagates the value to lacros so that their behavior can be
synchronized. This CL adds a kill switch in case we determine that it
has to be disabled. The functionality currently remains disabled.

When the functionality is enabled, the ash-based chrome app publisher
treat all ash-based chrome apps as disabled.

Change-Id: I194542049b5bbfa6746c60f4e01ab5733a78402a
Bug: 1225848
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198698
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927979}