Reland: [IntentHandling] Register Chrome app file_handlers
Reland this CL with some changes to make sure we publish the
QuickOffice extension in ExtensionAppsChromeOs. It has
file_handlers and was previously dealt with in
FindFileHandlerTasks, so we need to move it to App Service too.

CL:3189394 added a test for QuickOffice that fails under the
old CL without these changes.

This reland also adds some pieces from the follow up CL:3168581
which are actually necessary as there are tests in the tree
that rely on them but don't run on pre-submit.

Diff from original CL:
 - Allow QuickOffice in ExtensionAppsChromeOs::Accepts()
 - Exempt QuickOffice in ShouldIgnoreApp()
 - app_service_file_tasks.cc pieces from CL:3168581
 - CreateChromeAppIntentFilters() pieces from CL:3168581


Original description:

Create IntentFilters for Chrome app file_handlers that allow them
to show up in GetAppsForIntent() for Intents with files.

Add some functionality to the Intent system like matching against
directories, which is needed for the include_directories option
for Chrome app file_handlers which does not exist for SWAs/PWAs.

We add is_directory to IntentFile, rather than try to guess based
on URL, because we actually have this information in
FindAppServiceTasks() - the EntryInfo for each file has an
is_directory field, and this info is from the file system so is
much more trustworthy.

Change GetAppsForFiles() to accept a list of IntentFiles directly,
as we would otherwise need to add an array of is_directory bools
anyway, which is getting a bit cumbersome.

Move behaviour from IsGoodMatchFileHandler() in file_tasks.cc into
FindAppServiceTasks(). This deals with finding the "best" handler
within a single app for the given intent - based on product
decisions. We can't do this in post-processing because we only
return one handler per app. This allows us to return
is_generic_file_handler in each IntentLaunchInfo.

Ignore all Chrome app handlers with verbs other than "open_with".
In Files app, we were already ignoring these handlers and there
was no way to see them or use them from the UI. This CL extends
that into the FindAllTasks API, which requires changing a few
tests which expected handlers for other verbs to be returned.
This allows a simpler implementation and has no user-visible
effects.

Drive-by fix a bug in FilterIsForFileExtensions() where we didn't
account for kAction conditions.

Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Bug: 1240018
Change-Id: Ic2df3af1c7737d2ce4fe752c758711e488128b1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193930
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927980}