[FilesSWA] Record FSP mount UMA in event router when SWA enabled
The background page currently records
FileBrowser.FileSystemProviderMounted and in the SWA world the
background code will exist on every window. This ports the code to C++
specifically in the event_router.cc and guards it to only emit if the
SWA flag is enabled.

There is a special use case where the SWA flag might be enabled but the
Chrome app is still running (thus a background page still starts). In
this case, favour the event router implementation over the background
page.

Furthermore, the mount_metrics.js enum for FSPs had the ZIP_ARCHIVER and
ZIP_UNPACKER removed. When the JS API implementation fires first it sets
up the Histogram to contain the num of values in the array (i.e. 16) but
the C++ API requires explicit values (skipping them throws an error).
There is also guidance to keep the the
tools/metrics/histograms/enums.xml file up to date with the
implementation.

Bug: 1254243
Test: browser_tests --gtest_filter=*deduplicatedUmaMetricForFile*
Test: Compile, flash to device and install Dropbox FSP
Change-Id: I79ee60b7978843f558f9f6dd649c1dffe9eca5b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3195849
Reviewed-by: Bo Majewski <majewski@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#927997}