settings: Overwrite invalid input method prefs values.
Right now, it's possible for input method prefs to contain invalid
values (e.g. as the result of deprecations). The settings page will
fallback to a default value for that option, but the prefs will remain
invalid until the user selects a different value.

This is a bit fragile and could bite us in the future when extending
an option with new values that clash with deprecated values. So add a
mechanism in the input method settings code where we overwrite invalid
pref values with the default value.

Bug: 1239987
Change-Id: I8a891abbcf26067fdcddb25e9af62ed92a3e6077
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3192881
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: John Palmer <jopalmer@chromium.org>
Reviewed-by: Bao-Duy Tran <tranbaoduy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928016}