Don't inherit cross-BrowsingInstance proxies in new child frames.
Currently, when a parent frame creates a new subframe, we look through
the list of proxies in the parent's FrameTreeNode and create a set of
proxies in the same SiteInstances in the new subframe's FrameTreeNode.
This turns out to be problematic if the proxy is in a different
BrowsingInstance, which turns out to be possible if we have started a
main frame navigation to a different BrowsingInstance, and before it
commits, the current (old) document adds a new subframe.  In
particular, creating that proxy on a subframe FTN means that the proxy
may get preserved if the old document is stored and later restored
from BackForwardCache, leading to crashes when we try to use that
proxy while assuming it's in the same BrowsingInstance, such as if the
subframe creates another subframe.

More concretely, consider the following example.  Suppose A starts a
navigation to B.  This goes through
CreateProxiesForNewRenderFrameHost, which creates a
cross-BrowsingInstance proxy in the main frame's FTN, because all
speculative RFH navigations must always start with a proxy that will
be replaced by a real frame on commit.  Note that we intentionally
skip subframes of the navigating frame in this code, so while process
B will have a proxy for the main frame, it won't know about any
subframes that A might've had before the navigation to B started.

Next, suppose A adds a subframe A2 before the navigation to B commits.
Presently, this copies all proxies in A's FTN to the new A2 FTN,
*including* the cross-BrowsingInstance proxy for B.  This is
incorrect: the new process doesn't need to know about the new
subframe, just like it didn't need to know about A's subframes when
the navigation started.  There's still a safeguard in that if we allow
the proxy to be created, it would normally get cleaned up as part of
Frame::Swap(), when detaching the old frame's children when the
navigation commits, but unfortunately it's possible to bypass that
safeguard by racing the commit IPC with the proxy creation IPC, in
which case the cleanup won't happen.  In that case, A2's FTN would
keep the proxy in B's process, and if A went into BFCache and was
later restored, A2 would still have that proxy.  If A2 now creates
another subframe A3, we will try to inherit all proxies from A2 to A3,
but this will break a bunch of our assumptions (namely, that subframes
are in the same BrowsingInstance, and that a subframe proxy implies
that a corresponding RenderViewHost has already been created), which
leads to a crash.

To fix this, we modify CreateProxiesForChildFrame to skip over
proxies in a different BrowsingInstance.

Bug: 1243541
Change-Id: If9fe89302c5338ec83549e8827611e8e7f3c6525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3203096
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928045}