Use system slot client certificates for SAML flow on lock screen
The lock screen can show a SAML reauthentication dialog.
If this happens, allow usage of system slot client certificates there,
with the same restrictions as apply for system slot client certificate
usage in SAML flows on the sign-in screen.

Note that this does not add support for smart card provided certificates
on the lock screen - that is tracked in https://crbug.com/1255676 .

LockscreenReauthDialogTestHelper has been introduced to hold shared
functionality for testing the lock screen re-auth dialog. It is used
by existing SamlLockscreenBrowsertest and the new browsertest.

Bug: 1252678
Test: New browsertest
Change-Id: Ic12ba218de3f436a8f55d4f3873455e5123b9e53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193015
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928049}