[Lacros] Allow Lacros-extension set proxies in Ash
Proxies set in the Lacros primary profile should be propagated to Ash.
Ash will forward the extension set proxies to the PlayStore and
to Chrome OS system services.

The proxy will remain active in Ash until any extension in the primary
profile clears the proxy (the proxy will get cleared in Lacros and
Lacros will send a request to Ash to clear the proxy) or until the
extension which is controlling the proxy is uninstalled.

This change is required so we don't break expectations of existing
customers.

This CL was split from CL:3162599.

Bug: b:192915736
Test: browser_tests
Change-Id: I1604d48cfe090fccabaf92ce6acba1ef6998970d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193532
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Cr-Commit-Position: refs/heads/main@{#928088}