Fix widget inputs crashes in fenced frame.
The check for a containing render process host was not updated
for inner frame trees. It turns out that a check in
RenderFrameHostImpl::CreateNewPopupWidget already covers this
situation and is more accurate checking the frame tree state so
the problematic code can be removed.

BUG=1123606

Change-Id: I97f222e625a5466524586417aaa923deab91d22c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3193214
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928153}