Hats Next: Move settings triggers over to use in-built same-origin logic
Previous CLs shifted the ability to detect whether the user has remained
on the same origin during the survey delay onto the HaTS service, and
out of the HaTS handler.

This CL updates the HaTS handler to actually make use of this moved
logic.

Bug: 1249429
Change-Id: Ib5af4c2fdc7fc68d46f82a689d61c3033ec4b21f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3205201
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/main@{#928180}