Introduce temporary storage layer to Impression Attributions
When Impressions are received in the background, we don't want to
always start up the Browser to process them as this is very heavy
weight, so instead we persist them to disk.

This CL does not hook up the storage layer to the rest of the
pipeline yet - that will come in followup changes.

Raw files are used since data integrity is not vital to maintain. If
background attributions get lost (at reasonably low rates), it's not
going to cause user-visible issues. Attributions are appended linearly
to the file, so if any corruption or truncation happens, Attributions
that precede the truncation/corruption can still be processed.

Bug: 1210171
Change-Id: I4c14684cc5445b153218ea8cd5693c265957c213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3152895
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928231}