Preserve Surface frame timestamp attrib on resize
When a NativeViewGLSurfaceEGL is resized or recreated, the underlying
EGL surface is replaced with a fresh object. If SetEnableSwapTimestamps
was used on the old surface, the new surface should also have
EGL_TIMESTAMPS_ANDROID set to true.

This fix avoids validation errors in ANGLE which verifies that
the EGL_TIMESTAMPS_ANDROID property was set when retrieving
timestamps.

Bug: 1197111
Change-Id: I928d45d89bd8308916b45777d50a2cd493e23667
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199338
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928236}