[CrOS Bluetooth] Update DeviceNameManager to validate device_id exists.
Update DeviceNameManager to check if a device with identifier
|device_id| exists in the Bluetooth adapter before setting the device's
nickname. If no device does not exist, the nickname is not saved. This
is done because we have no hard-limit on number of device nicknames
that can be saved. In order to add some constraint, we are only allowing
devices that have been found to have nicknames set. This limits the
number of nicknames a user can set.

Bug: 1010321
Change-Id: Icdef8ff4f0725a9b77230792ee37bf23e402c833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3202576
Commit-Queue: Gordon Seto <gordonseto@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928251}