[Private Network Access] Disable secure context restriction on webview.
Per go/finch-killswitch#merge-policy, a kill switch must be followed up
with a merge to the affected channel (here webview stable) to disable
the feature by default.

ntfschr@ also points out that the kill switch might not work for the
devices affected by b/201865749, which might all run for the first time.

This CL makes use of the existing ContentBrowserClient interface to
selectively disable the deprecation on Webview only. Returning true
from ShouldAllowPrivateNetworkRequests results in the content/ code
ignoring the value of the BlockInsecurePrivateNetworkRequests Finch
feature flag. This means that even though the flag may be set to true
by default for webview, it is ignored and the deprecation is disabled.

Bug: chromium:1255675
Change-Id: I21c60acbea2bf3b7598b659ecbbac0f447fb5d1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3199763
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Auto-Submit: Titouan Rigoudy <titouan@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928260}