Fix Range loop in DisplayLockDocumentState
In http://crrev.com/925413 I fixed a very similar Range loop in
DisplayLockUtilities, but I forgot to update the other loop in
DisplayLockDocumentState. This patch updates the
DisplayLockDocumentState Range loop to match the newer
DisplayLockUtilities one.

This patch fixes the attached clusterfuzz test case, but I'm not
including a test case because the clusterfuzz test doesn't have a
minimal repro.

Ideally I would make these loops share the same code, but I couldn't
figure out an elegant way to do so. I could make a function that takes a
lambda, but the usage in DisplayLockDocumentState wants to return out of
the top level function in the middle of the loop, which I don't think I
could do from a lambda. I could try to make my own Iterator class for
this... thoughts?

Fixed: 1255155
Change-Id: I23025a52cd23b4a1a49b48def40fd016f42edbba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3205667
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928283}