[side search] Fix focus issues and auto-open behavior
This CL addresses current issues with focus in the side panel.
Specifically this CL allows the side panel to request focus when
opened via the toolbar button but not when it is opened as a
result of switching to a tab where the side panel is opened.

The SideSearchBrowserController uses an ExternalFocusTracker to keep
track of the last focused view and will restore focus to this view
when the side panel is closed. This fixes current behavior where
focus would move to the next available view in focus order when
closed (focus was always moved into the security chip).

This CL also prevents the side panel from auto-reopening when the
user has gone from
  non-Google page with side panel open
  Google SRP (or similar) where the panel auto closes
  another non-Google page where the side panel currently auto reopens

In the new flow (in the state-per-tab model) the side panel would not
reopen for the last navigation.

Fixes to propagation of user prefs and task manager ids have also
been bundled into this CL.

Bug: 1236631
Change-Id: I7ff7153d8f8c51b8278b62592270cd2eaf15c1d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3200397
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928287}