[WebLayer] Eliminate flakiness in subresource filter browsertest
The expectation verifies that there's a message in the console when the
subresource filter has been activated for a webpage. That expectation
has been failing on the bots, and I can repro the failure locally. The
problem is due to the fact that the test doesn't call Wait() on the
console observer but rather directly verifies that the console has a
message; this direct check races with the message appearing.

This CL changes the check to instead wait for the message to appear. I
have verified locally that the flake no longer presents itself.

Hat tip to csharrison@ for dianosing the problem :).

Bug: 1252428
Change-Id: I798be7c422e7ac3629aa041619d72189b22fedeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198883
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928300}