Show warning if needed on video capture stopped
In addition to checking whether ongoing video recordings should be
interrupted due to DLP restrictions, we also check if there was some
confidential content shown during the recording when the restriction
level is only WARN. In that case, we show a dialog at the end of the
recording informing the user about this and letting them choose whether
to save or delete the file.

The actual saving/deleting of the recording is not part of this change.

Bug: 1222814
Change-Id: I334baa87ad0add338de14b74b9614ffb45a193b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3194612
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928302}