Reland "Disable Window Controls Overlay when the Infobar is visible"
The test failure occured because UpdateWindowControlsOverlayEnabled
was getting called more times during navigation due to the animation of
the toolbar, and the draggable region was reset too early during the
navigation. This caused the webcontent to update the bounds of the
draggable region (due to the change in the overlay) which set a
non-empty value for the draggable region after it had been cleared.

The fix was to change the draggable region reset from
ReadyToCommitNavigation to DidFinishNavigation, so that there wouldn't
be any more draggable region updates from the old web contents after the
value was cleared.

crrev.com/c/3180900/1: Original change
crrev.com/c/3180900/2: A test "fix" that covered up the underlying
                       issue.
crrev.com/c/3180900/3: Reverting the "fix"
crrev.com/c/3180900/4: Actual fix that addressed the root cause of the
                       test failure.

> This reverts commit e6d488f84fa485f7c99ff0e116628fe620b6ef79.
>
> Reason for revert: Suspected of causing
> WebAppBrowserFrameViewMacWindowControlsOverlayTest.
> WindowControlsOverlayDraggableRegions (crbug.com/1252632)
>
> Original change's description:
> > Disable Window Controls Overlay when the Infobar is visible
> >
> > Currently, when Window Controls Overlay (WCO) is enabled and the infobar
> > is visible, the infobar flows under the overlay. This blocks the user
> > from being able to read the content of the infobar and prevents them from
> > being able to close the infobar with the mouse (it's still reachable via
> > keyboard).
> >
> > This change reverts the app back to a standalone title bar when the
> > infobar appears, and returns the app back to WCO once the infobar is
> > closed.
> >
> > Before and after: https://imgur.com/a/cx0Afk8
> >
> > Bug: 1229406
> > Change-Id: Ia7747e80b37ec5ec8811a921df06d3074b6ef6e6
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3146234
> > Commit-Queue: Amanda Baker <ambake@microsoft.com>
> > Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
> > Cr-Commit-Position: refs/heads/main@{#924520}

Bug: 1229406
Change-Id: Ie572021982093890cfff50a88d53c81deb40bdcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3180900
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Amanda Baker <ambake@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#928303}