arc: Show "Check network" button on sign-in failure
We show a button in optin dialog to check network when there is failure
that can potentially be due to network connectivity problem. Currently
this button is not shown when the optin flow fails with
GMS_SIGN_IN_NETWORK_ERROR error which is a clear indication of network
problem so we updated our check to include this error in the list of
errors that will show the button.

Bug: b:190714890
Test: unplug network during optin, button displays on error screen
Change-Id: I3ddcb9f4cd5d73dd976503baa0650d31c9630f45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3205399
Reviewed-by: Long Cheng <lgcheng@google.com>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928317}