Add new features to segmentation platform
Adds Dummy, ChromeStart and QueryTiles features to segmentation platform.
The features are not yet used, this CL adds the setup for the features
and updates the UMA histograms to work for multiple features.
Leaves the adaptive toolbar histograms while adding generic histograms
that will replace them.
Also moves the list of segmentation keys to components/ so it is easy
to record uma histograms

BUG=1251733

Change-Id: I9f0a44a6c4247c308a4e7967fa9ca4e5dce1d3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3179807
Commit-Queue: ssid <ssid@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928342}