StorageBuckets: New BucketLocator object
This change adds a new BucketLocator object to represent
a Storage Bucket. This includes a bucket's main identifying
properties (id, storage_key, type, is_default) to be passed between
Quota and Storage APIs. Unlike the BucketInfo object, all
included properties for the bucket should never change until
the bucket is deleted from the QuotaDatabase. This is expected
to be used in most cases where a Bucket is retrieved when
BucketInfo is too much but BucketId is not sufficient
(ex. QuotaClient, some methods on QuotaManagerProxy).

Change-Id: I42a0b99aaff51d6706350ee17d159c2b9860b491
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3183259
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#928373}